+++ pyxem/signals/common_diffraction.py
@@ -17,12 +17,19 @@
 # along with pyXem.  If not, see <http://www.gnu.org/licenses/>.
 
 import numpy as np
+import dask.array as da
 
 from hyperspy.api import interactive
 from hyperspy.misc.utils import isiterable
 import hyperspy.api as hs
 
 from traits.trait_base import Undefined
+from pyxem import CUPY_INSTALLED
+from hyperspy.misc.utils import _get_block_pattern, add_scalar_axis
+from tlz import concat
+
+if CUPY_INSTALLED:
+    import cupy as cp
 from pyxem.utils.virtual_images_utils import normalize_virtual_images
 
 
@@ -34,6 +41,54 @@ OUT_SIGNAL_AXES_DOCSTRING = """out_signal_axes : None, iterable of int or string
 
 class CommonDiffraction:
     """Common functions for all Diffraction Signals classes"""
+
+    def to_device(self):  # pragma: no cover
+        """Transfer the data to the GPU.
+
+        Notes
+        -----
+        This code is currently in an experimental stage. Please report and issues you have on the pyxem github
+        """
+        if not CUPY_INSTALLED:
+            raise ImportError(
+                "The cupy package is required to use this method. "
+                "Please install it using `conda install cupy`."
+            )
+        if not self._gpu:
+            if self._lazy:
+                self.data = self.data.map_blocks(
+                    cp.asarray
+                )  # pass chunk-wise the data to GPU
+            else:
+                self.data = cp.asarray(self.data)  # pass all the data to the GPU
+
+    def to_host(self):  # pragma: no cover
+        """Transfer the data from the GPU to the CPU.
+
+        Notes
+        -----
+        This code is currently in an experimental stage. Please report and issues you have on the pyxem github
+        """
+        if not CUPY_INSTALLED:
+            raise ImportError(
+                "The cupy package is required to use this method. "
+                "Please install it using `conda install cupy`."
+            )
+        if self._gpu:
+            if self._lazy:
+                self.data = self.data.map_blocks(cp.asnumpy)
+            else:
+                self.data = cp.asnumpy(self.data)
+
+    @property
+    def _gpu(self):
+        if not CUPY_INSTALLED:
+            return False
+        else:
+            if self._lazy:
+                return isinstance(self.data._meta, cp.ndarray)
+            else:
+                return isinstance(self.data, cp.ndarray)
 
     @property
     def unit(self):
@@ -243,3 +298,135 @@ class CommonDiffraction:
         if not self.metadata.has_item("Navigation_signals"):
             self.metadata.add_node("Navigation_signals")
         self.metadata.Navigation_signals.add_dictionary(dict_signal)
+
+    def _map_blocks_prepare(self, navigation_chunks="auto"):
+        """Prepare the signal for mapping blocks.  This function will check if the signal
+        is chunked correctly and rechunk if necessary.
+        """
+        if not self._lazy:
+            s_input = self.as_lazy()
+            s_input.rechunk(nav_chunks=navigation_chunks)
+        else:
+            s_input = self
+        chunk_span = np.equal(s_input.data.chunksize, s_input.data.shape)
+        chunk_span = [
+            chunk_span[i] for i in s_input.axes_manager.signal_indices_in_array
+        ]
+
+        if not all(chunk_span):
+            old_sig = s_input.rechunk(inplace=False, nav_chunks=None)
+        else:
+            old_sig = s_input
+        return old_sig
+
+    def _blockwise(
+        self,
+        func,
+        signal_shape=None,
+        navigation_shape=None,
+        dtype=None,
+        inplace=False,
+        lazy_output=None,
+        ragged=False,
+        num_workers=None,
+        meta=None,
+        **kwargs,
+    ):
+        """Apply a function to each block of the signal. This function might change in the
+        future.
+
+        Parameters
+        ----------
+        func : callable
+            The function to apply to each block.
+        *args : list
+            The arguments to pass to the function.
+        signal_shape : tuple
+            The shape of the signal.
+        dtype: np.dtype
+            The type of the signal.
+        **kwargs : dict
+            The keyword arguments to pass to the function.
+
+        Returns
+        -------
+        Diffraction2D
+            The signal with the function applied to each block.
+        """
+        if dtype is None:
+            dtype = self.data.dtype
+        if signal_shape is None:
+            signal_shape = self.axes_manager._signal_shape_in_array
+        if navigation_shape is None:
+            navigation_shape = self.axes_manager._navigation_shape_in_array
+        if lazy_output is None:
+            lazy_output = self._lazy
+        new_shape = navigation_shape + signal_shape
+        old_sig = self._map_blocks_prepare()
+        if meta is None:
+            meta = old_sig.data._meta
+
+        arg_pairs, adjust_chunks, new_axis, output_pattern = _get_block_pattern(
+            (old_sig.data,), new_shape
+        )
+        axes_changed = len(new_axis) != 0 or len(adjust_chunks) != 0
+
+        mapped = da.blockwise(
+            func,
+            output_pattern,
+            *concat(arg_pairs),
+            adjust_chunks=adjust_chunks,
+            new_axes=new_axis,
+            align_arrays=False,
+            dtype=dtype,
+            concatenate=True,
+            meta=meta,
+            **kwargs,
+        )
+        data_stored = False
+        if inplace:
+            if (
+                not self._lazy
+                and not lazy_output
+                and (mapped.shape == self.data.shape)
+                and (mapped.dtype == self.data.dtype)
+            ):
+                # da.store is used to avoid unnecessary amount of memory usage.
+                # By using it here, the contents in mapped is written directly to
+                # the existing NumPy array, avoiding a potential doubling of memory use.
+                da.store(
+                    mapped,
+                    self.data,
+                    dtype=mapped.dtype,
+                    compute=True,
+                    num_workers=num_workers,
+                )
+                data_stored = True
+            else:
+                self.data = mapped
+            sig = self
+        else:
+            sig = self._deepcopy_with_new_data(mapped)
+
+        am = sig.axes_manager
+        sig._lazy = lazy_output
+
+        if ragged:
+            axes_dicts = self.axes_manager._get_navigation_axes_dicts()
+            sig.axes_manager.__init__(axes_dicts)
+            sig.axes_manager._ragged = True
+        elif axes_changed:
+            am.remove(am.signal_axes[len(signal_shape) :])
+            for ind in range(len(signal_shape) - am.signal_dimension, 0, -1):
+                am._append_axis(size=signal_shape[-ind], navigate=False)
+
+        if not ragged:
+            sig.axes_manager._ragged = False
+            if signal_shape == () and am.navigation_dimension == 0:
+                add_scalar_axis(sig)
+            sig.get_dimensions_from_data()
+        sig._assign_subclass()
+
+        if not lazy_output and not data_stored:
+            sig.data = sig.data.compute(num_workers=num_workers)
+        return sig